+++ src/main/java/ru/practicum/shareit/booking/dto/BookingDto.java
@@ -13,6 +13,6 @@ public class BookingDto {
     private LocalDateTime start;
     private LocalDateTime end;
     private Integer itemId;
-    private Long bookerId;
+    private Integer bookerId;
     Status status;
 }
