+++ src/main/java/ru/practicum/shareit/user/dao/UserDao.java
@@ -3,6 +3,6 @@ package ru.practicum.shareit.user.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.shareit.user.model.User;
 
-public interface UserDao extends JpaRepository<User, Long> {
+public interface UserDao extends JpaRepository<User, Integer> {
 
 }
