+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -9,11 +9,11 @@ public interface UserService {
 
     List<UserDto> getAllUsers();
 
-    UserDto updateUser(UserDto userDto, long id);
+    UserDto updateUser(UserDto userDto, int id);
 
-    UserDto findUserById(long id);
+    UserDto findUserById(int id);
 
-    void deleteUserById(long id);
+    void deleteUserById(int id);
 
-    boolean isExist(long id);
+    boolean isExist(int id);
 }
